Display content in calculator all buttons work
It doesn't operate yet but it displays the content
well.

Bugs to solve:
1. When you have pressed the first number and then
the operator you have to be able to change the
operator by pressing another operation button.
But if you have pressed the second number when
you press the operator the outcome should be the
result plus the operator in the upper part
of the panel and the result in the lower part.

2. The = button only should be actionable when
the user has pressed a first number, an operator
and a second number.